Rename fields to match Wikidata
Rather than area / areaid and party / partyid, label the fields such
that the "main" column is the Wikidata item id, and the human-readable
text is a label. This will more closely match how we tend to get the
data back from the Query Service, and make it easier to generate diffs
between its output and ours.